test(core): split agent work state responsibilities
(cherry picked from commit 93c96d58f74aa78d7a1e481536c81622dd3cff71)
