+++ developer/sdk/kfd/kfd-2/claims/agent-work-state-contract.json
@@ -43,7 +43,7 @@
       "pointer": {
         "kind": "file",
         "path": "framework/core/tests/python/test_agent_work_state_contract.py",
-        "sha256": "243df94d5ce449aa3031f84981ebd79d867de39f3b11c84f31caa4490a3b4c5e"
+        "sha256": "6d160a1f10fc900dc3e7057394131dae5fa73d93902f47a0adf329e8792909d5"
       },
       "description": "Runtime CLI parity and fail-closed contract-schema checks."
     },
@@ -75,7 +75,7 @@
       },
       {
         "path": "framework/core/tests/python/test_agent_work_state_contract.py",
-        "sha256": "243df94d5ce449aa3031f84981ebd79d867de39f3b11c84f31caa4490a3b4c5e"
+        "sha256": "6d160a1f10fc900dc3e7057394131dae5fa73d93902f47a0adf329e8792909d5"
       },
       {
         "path": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
