+++ developer/sdk/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.5",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.5",
-    "sourceSha": "c791b22287be4399a867fbcfdd44f3202e5105ef"
+    "sourceSha": "8c3588ff5376ba63ae769d42008f5168c49ec568"
   },
   "claims": [
     {
@@ -260,7 +260,7 @@
           "pointer": {
             "kind": "file",
             "path": "framework/core/tests/python/test_agent_work_state_contract.py",
-            "sha256": "243df94d5ce449aa3031f84981ebd79d867de39f3b11c84f31caa4490a3b4c5e"
+            "sha256": "6d160a1f10fc900dc3e7057394131dae5fa73d93902f47a0adf329e8792909d5"
           },
           "description": "Runtime CLI parity and fail-closed contract-schema checks."
         },
