+++ framework/data-protection/work-agent-history.contract.json
@@ -148,7 +148,7 @@
     "nativeMigrationTests": [
       "framework/core/tests/python/test_agent_work_profile_native.py::test_native_profile_authority_bundle_restores_clean_home_exactly",
       "framework/core/tests/python/test_agent_work_profile_native.py::test_native_profile_authority_import_preflights_all_operations_before_writing",
-      "framework/core/tests/python/test_agent_work_state_contract.py::test_kfd7_legacy_role_roots_remain_readable_without_reinterpretation"
+      "framework/core/tests/python/_agent_work_state_role_cases.py::test_kfd7_legacy_role_roots_remain_readable_without_reinterpretation"
     ],
     "requiredSurfaces": [
       "source",
